update Hugo to the latest version, switch to Goldmark and  deprecate Blackfriday.
I have added markdown render hooks to open links in new tabs and
add an anchor link to all headings.

There is one issue: The render hooks are not fired for linkify
links. Blackfriday did not have this issue. For now, I have to
either replace all linkify links or let them be.